fix(TUP-38847):connection error popup with every change of the update url from Update settings https://jira.talendforge.org/browse/TUP-38847

+++ main/plugins/org.talend.repository/src/main/java/org/talend/repository/ui/login/connections/settings/UpdatesitePreferencePage.java
@@ -703,7 +703,7 @@ public class UpdatesitePreferencePage extends PreferencePage {
             @Override
             public void modifyText(ModifyEvent e) {
                 updateLocalBasicAuthUI();
-                setValid(validateBasicAuth());
+                setValid(true);
                 refresh();
             }
         });
@@ -712,7 +712,7 @@ public class UpdatesitePreferencePage extends PreferencePage {
             @Override
             public void modifyText(ModifyEvent e) {
                 updateLocalBasicAuthUI();
-                setValid(validateBasicAuth());
+                setValid(true);
                 refresh();
             }
         });
@@ -1162,6 +1162,11 @@ public class UpdatesitePreferencePage extends PreferencePage {
     @Override
     public boolean performOk() {
         if (this.isControlCreated()) {
+            if (!validateBasicAuth()) {
+                setValid(false);
+                return false;
+            }
+
             try {
                 if (m2Delete.getSelection() != p2Service.removeM2()) {
                     p2Service.saveRemoveM2(m2Delete.getSelection());
